pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "src.ttspod.version.__version__"}  # any module attribute compatible with ast.literal_eval

[project]
name = "ttspod"
dynamic = ["version"]
description = 'Convert any content to a podcast'
authors = [
  {name = "Adam Kessel", email = "adam@rosi-kessel.org"}
]
maintainers = [
  {name = "Adam Kessel", email = "adam@rosi-kessel.org"}
]
keywords = [ 'podcast', 'text-to-speech' ]
readme = 'README.md'
license = {text = "MIT License"}
classifiers=[
  'Development Status :: 3 - Alpha',
  'Intended Audience :: End Users/Desktop',
  'License :: OSI Approved :: MIT License',
  'Programming Language :: Python :: 3.11',
  'Operating System :: POSIX :: Linux',
  'Operating System :: MacOS :: MacOS X',
  'Environment :: MacOS X',
  'Environment :: Win32 (MS Windows)'
] 
requires-python = '==3.11.*'
# spacy lock on 3.7.5 to avoid conflicts
dependencies = [ 
  'anyascii',
  'cython',
  'fastapi',
  'Requests',
  'Unidecode',
  'ffmpeg',
  'html2text',
  'instapaper',
  'ipython',
  'lxml',
  'lxml_html_clean',
  'nltk',
  'numpy<2',
  'pandoc',
  'paramiko',
  'pip',
  'pocket',
  'pocket_api',
  'pod2gen',
  'posix_ipc; platform_system != "Windows"',
  'psutil',
  'python-magic-bin; platform_system == "Windows"',
  'semaphore_win_ctypes; platform_system == "Windows"',
  'spacy==3.7.5',
  'pydub',
  'pymupdf',
  'pypandoc',
  'pypandoc-binary',
  'python-dotenv',
  'python_magic',
  'trafilatura',
  'validators'
  ]
[project.optional-dependencies]
local = ['torch>=2.2.2','torchaudio>=2.2.2','whisperspeech','webdataset','coqui-tts==0.24.2','coqui-tts-trainer==0.1.4']
remote = ['openai','elevenlabs']
truststore = ['truststore']
dev = ['posix_ipc','semaphore_win_ctypes','build','twine']

[project.urls]
Homepage = "https://github.com/ajkessel/ttspod"
Documentation = "https://github.com/ajkessel/ttspod/blob/main/README.md"
Repository = "https://github.com/ajkessel/ttspod"
Issues = "https://github.com/ajkessel/ttspod/issues"
Changelog = "https://github.com/ajkessel/ttspod/commits/main/"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = [ "src", "examples" ]

[tool.setuptools.package-data]
src = [ "*.py" ]
examples = ["dotenv.env"]

[project.scripts]
"ttspod" = "ttspod.cli:main"
"generate_samples" = "ttspod.generate_samples:main"
